Feature/issue 94 update teams goals (#96)
* Added update functionality for goals and teams

* changed object id to string

* created custom json converter for ObjectId

* Added a try catch block to the updating of teams in case of errors

+++ pokematic-backend/Models/Goal.cs
@@ -16,7 +16,7 @@ namespace pokematic_backend.Models
 
         [BsonElement("experiencePoints")] public int ExperiencePoints { get; set; }
 
-        [BsonElement("progress")] public int Progress { get; set; }
+        [BsonElement("progress")] public double Progress { get; set; }
 
 
     }
